updated type for reps

+++ FitTrack - SQFlite/fittrack/lib/models/exercises/ExerciseSet.dart
@@ -1,5 +1,5 @@
 class ExerciseSet {
-  double reps;
+  int reps;
   double weight;
 
   ExerciseSet({this.reps, this.weight});
